Add wlrs licence id to the nald licence import
https://eaflood.atlassian.net/browse/WATER-4545

Relevant changes - https://github.com/DEFRA/water-abstraction-system/pull/1294

As part of the ongoing work to migrate the legacy import jobs we need to establish if a licence already exists in the wrls database.

We check if the licence already exists, if it does then we want to do nothing. If it does not exist we want to generate return logs.

This change is added to highlight where we need to add this additional logic. And the service call will be added in another piece of work.