Create seeder for `charge_categories` (#1312)
Currently, with the way data is being inserted into `charge_categories` using the helper to create records during unit testing. We are getting intermittent failures when running the unit tests when the charge `reference` gets duplicated due to its lack of randomness.

We have therefore decided that since the data in `charge_categories` is reference data we shouldn't be creating new charge category records each time a unit test requires one. Instead, we should just be seeding the data once and then writing the unit tests to use this seeded data. The bulk of this work has been done for other tables in this PR https://github.com/DEFRA/water-abstraction-system/pull/1230

This PR will create the functions required to seed the `charge_categories` and then fix any code affected by this change.