Remove import jobs stats from /health/info
https://eaflood.atlassian.net/browse/WATER-4670

After making some changes to [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) we found it was no longer completing when run. The issue seemed to be

- a step we introduced to let [water-abstraction-system](https://github.com/DEFRA/water-abstraction) see importing licences was overloading the environment because it was acting like a DOS on the system app
- each job was taking much longer than realised to complete, and the existing job schedule was overlapping them, again leading to limited system resources

Initially, we were going to create a new 'job' that would daisy chain the existing jobs to avoid the overlap. Plus we'd drop the step that pinged the **water-abstraction-system** app. But then we realised based on current timings the whole process risked still running when users logged on in the AM.

So, we've always suspected that the features the previous team added to improve performance (pg-boss message queues and caching to redis) were actually impacting it. We've now gone through all the 'jobs' that were involved in importing data from NALD and stripped out this overhead. Our timings are much, _much_ better proving our suspicions.

What this means though is the block of job stats we display in the `/health/info` page is now redundant. In fact, with the jobs not being there when the update **water-abstraction-import** ships this page will error because the attempt to retrieve them will fail.

So, this strips out all logic related to job info from the `/health/info` page.