Update helpers to use seeded IDs where possible (#1339)
https://eaflood.atlassian.net/browse/WATER-4677

Whilst working on the ticket to remove the `DatabaseSupport.clean()` function from our unit tests we found that the licence helper had been using randomly generated UUIDs for the region ID which was causing some tests to fail.

This was fine while we were cleaning the DB after each test. But with the DB cleaning removed. Because some tests use the licence helper to add a licence, not caring about the region. These licences with fake region IDs are causing some of the other unit tests to fail when they expect an actual region to be associated with a licence.

So in this PR, the licence helper is going to be updated to use an actual random region ID. Which is now possible since the region data is seeded.

To keep things consistent we are also going to ensure that all other helpers that use a region ID are using a real region ID.

But why stop there :grin: Where we now have seed data, e.g. for the charge categories. We are going to update any helpers that have fake IDs for data that is seeded with real IDs.