https://eaflood.atlassian.net/browse/WATER-4579 (#1341)
* https://eaflood.atlassian.net/browse/WATER-4579

We recently started seeding the test DB with reference data. We then Restructured that seeding to be Knex based.

We thought we'd covered all reference data with the seeding. But we've just spotted that we missed the licence roles.

This change adds seeding for licence role. This has been done similar to the region where were there are no unique constraints on the table columns so we need to check if the licence role already exists and either patch or insert.