Update Licence role helper to use reference data
As part of ongoing work to seed reference data we have been updating the helpers to use the data we seed.

This change will update the licence role helper to select the data for the licence role.

It will default to licenceHolder and allow the default to be overridden when needed. This has been done as the majority of the current test use the licence holder role.