Add external id to addresses view. (#1346)
* Add external id to addresses view.

https://eaflood.atlassian.net/browse/WATER-4666

As part of the ongoing to work to migrate the import service into the water-abstraction-system, we need to add the external id for the addresses table.

When we import the addresses we need to upsert the data. The current import service uses the external id to resolve the conflict.

This follows the same approach as the other import work.