Add pre sroc returns flagging to service (#1335)
https://eaflood.atlassian.net/browse/WATER-4199
https://eaflood.atlassian.net/browse/WATER-4676

During testing for the first ticket, it was discovered that editing a return on a licence and then flagging that licence for supplementary billing, was always flagging the licence for pre-sroc supplementary billing even when the years were for sroc.

After discussions it was determined that editing a return should only be able to add a flag on a licence for either pre-sroc billing OR two-part tariff supplementary billing.
This is because regular sroc supplementary billing doesn't look at the returns, so editing a return won't change anything on the bill run.
This PR adds an additional check on the determine-return-log-years service to check if the return has a start date before the SROC start date. If it does we add the flag on the licence for pre-sroc supplementary billing.