Phase 2 of changes
I was hoping to separate changes at the data layer with those at the actual services and features.

However, no matter which way I go it looks like I'll be breaking existing tests. I could fix them, but I'm only going to then change them again in phase 2.

So, contrary to what I preach, this is going to have to be a big PR doing multiple things in order to get return requirements working again in a timely manner.

😳😬