Housekeeping - first pass of points validator tidy
The `SubmitPointsService` already handles dealing with the issue that where a single option is selected we don't get an array in the request payload.

So, what the validator was doing was unnecessary and never called. And as the calling service has already formatted the points on the payload, it might as well directly pass them in rather than the whole payload.

We've then updated the tests to reflect our changes but also to 'test' what the current tests claim to. In this commit the second test is failing not because it is mistakenly validating the data, but because we are getting a different error message (we had never covered the value not being a string).