Housekeeping - 3rd pass of points validator tidy
We actually know we are using the point ID as our value so we should be expecting and validating that we are getting a UUID.

So, the second test is failing again because we've changed the point to be the old NALD point ID which we have moved away from. As it is a string, no validation error is raised when now we expect one.