Fix show unregistered label in view licence page
https://eaflood.atlassian.net/browse/WATER-4683

It has been spotted when a licence has a registered user, in some case we are still displaying 'Unregistered licence' as the sub-heading above the title.

We've tracked it down to licences that have a registered user, but where an alias for the licence has not been added. We're not handling that scenario so falling back to displaying 'Unregistered licence'.

The fix is just to leave it blank.