Alter 2PT agreement logic in rtn req setup journey
https://eaflood.atlassian.net/browse/WATER-4672

> Part of the work to migrate return versions from NALD to WRLS

Currently, when creating a new return version if you select the 'Start with abstraction data' option the service will generate the return version based on data held against the licence. For each return requirement generated, we mark it as having a two-part tariff agreement if the licence has a two-part tariff agreement in place.

However, UAT has found this is too 'blunt' an approach. We're marking return requirements as two-part tariff even when the purpose linked to them is not.

So, in this change we drop looking at agreements against the licence. Now, we only mark the requirement as having a two-part tariff agreement if the purpose it is linked to is also flagged as two-part tariff. And this only applies when generating the return version from abstraction data against the licence.