Remove check for licence ended in approve rtn vers (#1356)
https://eaflood.atlassian.net/browse/WATER-4600

> Part of the work to migrate return versions from NALD to WRLS

We had to make a series of changes in all our projects to move away from using points data held in the `permit.licence` record's JSONB field when we found that if the licence was 'ended' (dead) [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) does not populate the `current_version` property. This is the only source for points data in the legacy solution.

Then we realised even though you could complete the journey, you'd get an error when you tried to approve the new return version! That is because of [WATER-4264](https://eaflood.atlassian.net/browse/WATER-4264), which included the following A/C.

```text
AC5: Check answers for the return requirements (error message for revoked/expired/lapsed licence)
Given I am viewing the 'Check your answers' page
And all the requirements have been entered correctly
When I select the 'Approve returns requirements' button but the licence is revoked, lapsed or expired
Then I will be shown the standard error page.
```

🤣

So, the final step to support adding return versions for ended licences is to remove the validation check we added to the approval process.