Fix show unregistered label in view licence page (#1357)
https://eaflood.atlassian.net/browse/WATER-4683

It has been spotted that when a licence has a registered user, we sometimes display 'Unregistered licence' as the subheading above the title.

We've tracked it down to licences with a registered user, but no alias was added. We're not handling that scenario, so we're falling back on displaying 'Unregistered licence'.

The fix is to leave it blank.