Add site description to rtn. req. bookmark links (#1358)
https://eaflood.atlassian.net/browse/WATER-4685

> Part of the work to migrate return versions from NALD to WRLS

During the UAT of the new view return version and requirements page, users let us know that, though useful, the bookmark links that currently display a return requirement's reference would be even better if they included the site description as well.

It's not so bad for return versions with 2 or 3 return requirements. But with more than that, it can be harder to find or recall which requirement you want to return to.

Well, our users' wish is our command!