Alter 2PT agreement logic in rtn req setup journey (#1359)
https://eaflood.atlassian.net/browse/WATER-4672

> Part of the work to migrate return versions from NALD to WRLS

When creating a new return version, if you select the 'Start with abstraction data' option, the service will generate the return version based on data held against the licence. If the licence has a two-part tariff agreement in place, we mark each return requirement generated as having a two-part tariff agreement.

However, UAT has found this approach too 'blunt'. We're marking return requirements as two-part tariffs even when the purpose linked to them is not.

Now, we only mark the requirement as having a two-part tariff agreement if the purpose it is linked to is also flagged as two-part tariff. This only applies when generating the return version from abstraction data against the licence.

---------

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>