Remove `DatabaseSupport.clean()` from unit tests (#1327)
https://eaflood.atlassian.net/browse/WATER-4677

We currently have some unit tests that fail if `DatabaseSupport.clean()` is not run before each test. This is usually due to duplicate data being added to the DB causing a constraint violation.

Since we have introduced "seeders" to populate some of our reference data. It should now be possible to fix any tests that currently fail without the DB clean. This will speed up the running of the unit tests as it won't have to rebuild the DB after running some of them.