Remove import jobs stats from /health/info (#1336)
https://eaflood.atlassian.net/browse/WATER-4670

After making some changes to [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) we found it was no longer completing when run. The issue seemed to be

- a step we introduced to let [water-abstraction-system](https://github.com/DEFRA/water-abstraction) see importing licences was overloading the environment because it was acting like a DOS on the system app
- each job was taking much longer than realised to complete, and the existing job schedule was overlapping them, again leading to limited system resources

We've always suspected that the features the previous team added to improve performance (pg-boss message queues and caching to Redis) were actually impacting it.

Initially, we planned to go through all the 'jobs' involved in importing data from NALD and strip out this overhead. We [carried out a spike](https://github.com/DEFRA/water-abstraction-import/pull/1023), and our timings are much, _much_ better, proving our suspicions.

However, this was a live issue, so we removed the step and tweaked the times to get the import working again.

In the meantime, we'll either complete the spike or [WATER-4535, a replacement for the licence import](https://eaflood.atlassian.net/browse/WATER-4535). Either way, the block of job stats we display on the `/health/info` page will be redundant. In fact, we don't use or look at it anyway, even when there is an issue.

So, this removes all logic related to job information from the `/health/info` page in readiness.