Update view summary service test
The current test was claiming that the `water` schema was returning no licence versions but `permit.licence` did contain a `current_version` property.

So, the test was unrealistic. But sticking with the idea we are dealing with a 'broken' licence (it has no licence versions) now we are not looking at the `permit.licence` record you would expect to see no points or source returned.