Stop generating optional data in company helper
https://eaflood.atlassian.net/browse/WATER-4600
https://eaflood.atlassian.net/browse/WATER-4645

We spotted when working on [Updating all scenarios to use new points solution](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/137) in **water-abstraction-acceptance-tests** that we were getting intermittent failures.

When we looked at the logs we often saw

```text
[15:57:07.633] ERROR (12080): insert into "companies" ("company_number", "external_id", "id", "name", "organisation_type", "type") values ($1, $2, $3, $4, $5, $6) returning * - duplicate key value violates unique constraint "unique_external_id"
    err: {
      "type": "UniqueViolationError",
      "message": "insert into \"companies\" (\"company_number\", \"external_id\", \"id\", \"name\", \"organisation_type\", \"type\") values ($1, $2, $3, $4, $5, $6) returning * - duplicate key value violates unique constraint \"unique_external_id\"",
```

When we checked the `CompanyHelper` we spotted that we were populating fields, including `external_id` which are not actually required. The idea of the helpers is to populate the minimum needed, any thing is on the caller to provide. If we don't need to populate these fields, we stand a better chance of avoiding failures due to unique constraint clashes.