Add abstraction method to lic. purpose points
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We recently [Updated view licence summary to use new points data](https://github.com/DEFRA/water-abstraction-system/pull/1316). This is part of a series of changes to help us move away from extracting point information from the licence JSON blob in `permit.licence`.

We're also looking to migrate the pages linked to from the view licence summary, which are currently in the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui). One of them is information on 'points'. We thought we'd captured everything 'point' related in [the changes we made to the points schema](https://github.com/DEFRA/water-abstraction-service/pull/2638) but have since spotted 'Means of abstraction'.

This is linked to an abstraction purpose point in NALD and gives how water will be abstracted there, for example, "Surface Mounted Pump (Fixed)". We'll [populate this field](https://github.com/DEFRA/water-abstraction-import/pull/1029) as we import a licence's point information from NALD.

This change updates the view so we can access the new field.