Fix down migrations for return logs view (#1377)
https://eaflood.atlassian.net/browse/WATER-4542

> Part of the work to migrate return versions from NALD to WRLS

We've been working on generating return logs in this repo, taking over from what NALD does.

This has meant changing what we see in our `return_logs` view.

When preparing a release for `production`, we needed to move our local environments back to the version of **water-abstraction-system** to match production.

This also meant rolling back changes to views to be in sync. When we did that, though, we started hitting lots of errors.

We tracked it down to the `down()` functions in recent alter `return_logs` migrations being wrong. Essentially, they were resetting the view to its original state rather than the previous state.

Fixing the functions means if anyone has to do the same and return to this point in the future, running `npm run rollback` shouldn't break anything.