Stop adding link on due returns for unauth users
https://eaflood.atlassian.net/browse/WATER-4444

We have built a replacement view licence page, which includes displaying return logs to all users.

All users should be able to see return logs. But when we reviewed the legacy code we overlooked that only permitted users should see a link on 'due/overdue' returns that allows them to edit it.

The issue was reported on behalf of users with basic access. When they were clicking the link they were seeing an error. We checked what the legacy page does and it simply doesn't display a link.

So, this change updates our version to replicate this behaviour.