Migrate abstraction details view page
As the legacy licence view page has now been rebuilt, the next process is to update some of the links pointing to legacy pages.
One of these legacy pages is the abstraction details view. If we are able to move this over to our system, it will help towards removing the use of old legacy code.

This PR is focused on rebuilding the abstraction details view page from the legacy code base.