Stop adding link on due returns for unauth users (#1378)
https://eaflood.atlassian.net/browse/WATER-4700

We have built a replacement view licence page, which includes displaying return logs to all users.

All users should be able to see return logs. However, when we reviewed the legacy code, we overlooked that only permitted users should see a link on 'due/overdue' returns that allows them to edit them.

The issue was reported on behalf of users with basic access. When they clicked the link, they saw an error. We checked what the legacy page does, and it simply doesn't display a link.

So, this change updates our version to replicate this behaviour.