Handle invalid points better in view lic. summary (#1374)
https://eaflood.atlassian.net/browse/WATER-4600
https://eaflood.atlassian.net/browse/WATER-4645

In [a recent change](https://github.com/DEFRA/water-abstraction-system/pull/1316) we updated the view licence summary to use our new 'point' tables rather than the legacy `permit.licence` JSON blob.

In the change, we assumed that we would not get licences with licence version purposes that do not have points. We should have known better when dealing with NALD data!

In fairness, it is common for data to be corrupted from deletions happening in NALD that don't occur on our side. But when it happens, it means the page 'errors' rather than falling back to just showing nothing. Specifically, it's when extracting the 'Source of supply'.

So, in this change, we amend the presenter to avoid errors when dealing with 'bad' licence data.