Handle empty site description in returns setup
https://eaflood.atlassian.net/browse/WATER-4701

When working on WATER-4685 our QA team spotted that there are some return requirements where 'Site description' is not set.

We [made a change to handle this when viewing a return version](https://github.com/DEFRA/water-abstraction-system/pull/1380)  (we display nothing!)

But same issue applies when these return requirements are used for 'copied from existing' when setting up new return versions. They also display as null there.

However, their is a greater problem. If a user were to complete the manual journey, 'Site description' would be required. We don't want to create new records where a required field is empty.

The agreed solution in this scenario is to fall back to what we do when the user instead selects 'start with abstraction data' in the journey. There, we default site description to be the description of the first point against the licence version purpose.

So, in this change we'll add a fall back to use the first point's description if site description is null on the return requirement we're copying.