Refactor the fetch into own service
Queries of this size we have previously placed into their own service. We also know now it makes testing easier with all our database clean calls removed.

But as we now need to provide different source data in order to confirm our changes for site description work, this seemed the perfect opportunity to do it.