Stop display null site desc. in view licence rtn (#1383)
https://eaflood.atlassian.net/browse/WATER-4702

If a return log was created from a return requirement with no site description, when a user sees the return log listed in the view licence returns page, they will see `null`.

We've agreed that we should display nothing when this is the case. This change updates the view licence returns page to handle this scenario.