Handle empty site description in returns setup (#1384)
https://eaflood.atlassian.net/browse/WATER-4701

When working on WATER-4685, our QA team spotted that some return requirements do not have 'Site description' set.

We [made a change to handle this when viewing a return version](https://github.com/DEFRA/water-abstraction-system/pull/1380) (we display nothing!)

But the same issue applies when these return requirements are used for 'copied from existing' when setting up new return versions. They also display as null there.

However, there is a more significant problem. If a user were to complete the manual journey, 'Site description' would be required. We don't want to create new records with an empty required field.

The agreed solution in this scenario is to fall back to what we do when the user selects 'start with abstraction data' instead in the journey. There, we default the site description as the description of the first point against the licence version purpose.

So, in this change, we'll add a fallback to use the first point's description if the site description is null on the return requirement we're copying.

---

During this change we took the decision to refactor the fetch into its own service.

We have previously placed queries of this size into their own service. Now, we also know that it makes testing easier with all our database clean calls removed.