Flag licences removed from workflow for supp billing
https://eaflood.atlassian.net/browse/WATER-4137

As part of the work to flag licences for the new two-part tariff supplementary billing, we need to be able to flag licences that are removed from the workflow. If a licence is in workflow it does not get picked up for annual billing. This means if the licence is removed without any new charge information added, then we risk the licence not being billed for the year. During the deletion of the workflow record, the UI repo now makes a call to our supplementary endpoint to allow the system to flag the licence. This change adds an additional service to our ProcessBillingFlagsService, that can take the workflowId and work out if the licence needs to be flagged for two-part tariff supplementary billing.